// modules/transaction-manager/modules/signature-tracking/enum/st-navigate-enum.ts
export enum StNavigateProtectedParameter {
  SIGNATURES = 'signature_tracking_signatures',
  TRANSACTION = 'signature_tracking_transaction',
}

export enum EUsePaymentOFfProvidersService {
  ACH = '3',
}

export enum ESignatureTrackingUrlFlow {
  HOME = '/transaction-manager/signature-tracking',
  SIGNATURES = '/transaction-manager/signature-tracking/signatures',
  TRANSACTION = '/transaction-manager/signature-tracking/transaction',
}