// models/user-info.interface.ts
export interface IUserInfo {
  connectionDate: string;
  customerCode: string;
  duration: string;
  email: string;
  firstLastname: string;
  firstName: string;
  ip: string;
  lastConnectionDate: string;
  profile: string;
  secondLastname: string;
  secondName: string;
  signatureType: string;
  status: boolean;
  thirdLastname: string;
  userType: string;
  username: string;
}